requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.4.0
requests==2.27.1
urllib3==1.26.8
